remove old nginx/fulltest directory and rename fulltest-provider to fulltest
