--- nginx/fulltest-provider/success.htm
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>Open Quantum Safe interop test server for quantum-safe cryptography</title>
-</head>
-<body>
-<h1 align=center>
-Successfully connected using
-<!--# echo var="oqs_sig_name" default="unknown" -->-<!--# echo var="ssl_curve" default="unknown" -->!
-</h1>
-
-Client-side KEM algorithm(s) indicated:
-<!--# echo var="ssl_curves" default="no" -->
-</body>
-</html>
-
