+++ nginx/fulltest/build_ubuntu.sh
