+++ nginx/fulltest/ext-csr.conf
@@ -6,9 +6,22 @@ prompt = no
 [req_distinguished_name]
 CN = test.openquantumsafe.org
 
+[v3_intermediate_ca]
+basicConstraints = critical, CA:true, pathlen:0
+keyUsage = critical, digitalSignature, cRLSign, keyCertSign
+subjectKeyIdentifier = hash
+authorityKeyIdentifier = keyid:always
+certificatePolicies = 2.5.29.32.0
+extendedKeyUsage = clientAuth, serverAuth
+
 [v3_req]
-keyUsage = nonRepudiation, digitalSignature, keyEncipherment
+basicConstraints = critical, CA:FALSE
+keyUsage = critical, digitalSignature
 extendedKeyUsage = serverAuth
+subjectKeyIdentifier = hash
+authorityKeyIdentifier = keyid:always
 subjectAltName = @alt_names
+certificatePolicies=2.23.140.1.2.1
+
 [alt_names]
 DNS.1 = test.openquantumsafe.org
