+++ nginx/fulltest/ngx_event_openssl.patch
