+++ nginx/fulltest/success.htm
@@ -6,10 +6,11 @@
 <body>
 <h1 align=center>
 Successfully connected using
-<!--# echo var="oqs_alg_name" default="unknown" -->
-!
+<!--# echo var="oqs_sig_name" default="unknown" -->-<!--# echo var="ssl_curve" default="unknown" -->!
 </h1>
 
 Client-side KEM algorithm(s) indicated:
 <!--# echo var="ssl_curves" default="no" -->
 </body>
+</html>
+
